Refactor transport_test.go
